+++ frontend/styles/Signin.module.css
@@ -1,44 +1,64 @@
- /* signinbox 전체 틀 */
-.signinbox {
-    border: 1px solid rgb(49, 46, 46);
-    padding: 10px;
-    border-radius: 10px;
-    box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.2);
-    background-color: #fafafa;
-    max-width: 20rem;
-    margin: auto;
-    margin-top: 20px;
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 70vh;
+    background-color: #ffffff;
   }
 
- /* 입력칸 */
-.input{
-    border: 1px solid rgb(181, 183, 182);
-    width: 298px;
-    height: 30px;
-    border-radius: 5px;
+/* signinbox 전체 틀 */
+.signinbox {
+    display: flex;
+    flex-direction: column;
+    width: 300px;
+    padding: 2em;
+    border-radius: 10px;
+    background-color: white;
+    border: 1px solid #000; 
+    box-shadow: 0px 0px 5px 2px rgba(0, 0, 0, 0.1);
+  }
+
+.inputGroup {
+    display: flex;
+    flex-direction: column;
+    margin-bottom: 1em;
 }
 
-/* sign in 클릭 버튼 */
-.submitButton{
+.inputGroup label {
+    margin-bottom: 0.5em;
+  }
+  
+  .inputGroup input,
+  .inputGroup select {
+    padding: 0.5em;
+    border: 1px solid #ddd;
+    border-radius: 5px;
+    outline: none;
+  }
+  
+  .inputGroup input:focus,
+  .inputGroup select:focus {
+    border-color: #0070f3;
+  }
+  
+  .submitButton {
+    padding: 0.5em 1em;
     background-color: #0070f3;
     color: white;
-    width: 298px;
-    height: 30px;
-    border-radius: 5px;
     border: none;
+    border-radius: 5px;
     cursor: pointer;
-}
-
-/* 글자 중간에 배치 & 흰색 */
-.mid{
-    text-align: center;
-    color: white;
-}
-
-/* 줄 사이 공백 */
-.space{
-    height: 12px;
-}
+    transition: background-color 0.2s ease;
+  }
+  
+  .submitButton:hover {
+    background-color: #0051a2;
+  }
+  
+  .submitButton:active {
+    background-color: #003d7a;
+  }
 
 /* 계정 만드는 박스 */
 .account{
@@ -48,9 +68,9 @@
     border-radius: 10px;
     box-shadow: 0px 0px 10px 0px rgba(0, 0, 0, 0.2);
     background-color: #fafafa;
-    max-width: 20rem;
+    max-width: 300px;
     margin: auto;
-    margin-top: 20px;
+    margin-bottom: 100px;
 }
 
 /* 아이콘 & 글자 배치 */
